feat: refactor product and user hooks to use new admin types
- Updated `useProduct` and `useUser` hooks to import `CreateProductData` and `UpdateUserData` from the new `adminTypes` file.
- Added `CreateProductData` and `UpdateProductData` interfaces to define product creation and update structures.
- Introduced `OrderDetails` interface in `orderTypes` for better order management.
- Created `getStatusBadge` utility function for consistent order status styling.
- Refactored import paths for improved organization and clarity.
