+++ features/admin/hooks/useProduct.tsx
@@ -1,4 +1,3 @@
-import { CreateProductData, UpdateProductData } from "@/types";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { toast } from "sonner";
 
@@ -10,6 +9,8 @@ import {
   toggleStock,
   updateProduct,
 } from "@/features/admin/lib/productApi";
+
+import { CreateProductData, UpdateProductData } from "../types/adminTypes";
 
 export const useProduct = () => {
   return useQuery({
